pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "kage"
version = "0.1.0"
description = "AI-powered penetration testing CLI assistant"
readme = "README.md"
license = "MIT"
requires-python = ">=3.10"
authors = [
    { name = "Kage Contributors" }
]
keywords = [
    "security",
    "penetration-testing",
    "bug-bounty",
    "ai",
    "cli",
    "red-team",
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "Intended Audience :: Information Technology",
    "License :: OSI Approved :: MIT License",
    "Operating System :: POSIX :: Linux",
    "Operating System :: Microsoft :: Windows",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Security",
    "Typing :: Typed",
]

dependencies = [
    "typer>=0.9.0",
    "rich>=13.0.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "httpx>=0.25.0",
    "httpx-sse>=0.4.0",
    "pyyaml>=6.0",
    "jinja2>=3.1.0",
    "aiofiles>=23.0.0",
    "asyncio-atexit>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.0.0",
    "ruff>=0.1.0",
    "mypy>=1.0.0",
    "types-pyyaml>=6.0.0",
    "pre-commit>=3.0.0",
]
pdf = [
    "weasyprint>=60.0",
]

[project.scripts]
kage = "kage.cli.app:main"

[project.urls]
Homepage = "https://github.com/kage-security/kage"
Documentation = "https://github.com/kage-security/kage#readme"
Repository = "https://github.com/kage-security/kage"
Issues = "https://github.com/kage-security/kage/issues"

[tool.hatch.build.targets.sdist]
include = [
    "/src",
    "/templates",
    "/plugins",
]

[tool.hatch.build.targets.wheel]
packages = ["src/kage"]

[tool.ruff]
target-version = "py310"
line-length = 100
src = ["src"]

[tool.ruff.lint]
select = [
    "E",      # pycodestyle errors
    "W",      # pycodestyle warnings
    "F",      # pyflakes
    "I",      # isort
    "B",      # flake8-bugbear
    "C4",     # flake8-comprehensions
    "UP",     # pyupgrade
    "ARG",    # flake8-unused-arguments
    "SIM",    # flake8-simplify
]
ignore = [
    "E501",   # line too long (handled by formatter)
    "B008",   # do not perform function calls in argument defaults
]

[tool.ruff.lint.isort]
known-first-party = ["kage"]

[tool.mypy]
python_version = "3.10"
strict = true
warn_return_any = true
warn_unused_ignores = true
disallow_untyped_defs = true
plugins = ["pydantic.mypy"]

[[tool.mypy.overrides]]
module = "tests.*"
disallow_untyped_defs = false

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
addopts = "-ra -q"

[tool.coverage.run]
source = ["src/kage"]
branch = true

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "if TYPE_CHECKING:",
    "raise NotImplementedError",
]
